improve saving history - save the history id for this scoreboard activity so when we save again, its overwritten instead of saving a new entry

+++ app/src/main/java/com/dgnt/quickScoreboardCreator/data/history/entity/HistoryEntity.kt
@@ -8,7 +8,7 @@ import org.joda.time.DateTime
 
 @Entity
 data class HistoryEntity(
-    @PrimaryKey val id: Int? = null,
+    @PrimaryKey val id: Long? = null,
     val title: String,
     val icon: ScoreboardIcon,
     val lastLookedAt: DateTime,
